+++ src/chrono_vehicle/wheeled_vehicle/tire/ANCFTire.cpp
@@ -91,8 +91,8 @@ void ANCFTire::ProcessJSON(const rapidjson::Document& d) {
     // Read layer information for the Bead Section
     m_bead.num_divs = d["Bead Section"]["Number Elements"].GetInt();
     m_bead.num_layers = d["Bead Section"]["Layer Thickness"].Size();
-    assert(d["Bead Section"]["Ply Angle"].Size() == m_bead.num_layers);
-    assert(d["Bead Section"]["Material ID"].Size() == m_bead.num_layers);
+    assert(d["Bead Section"]["Ply Angle"].Size() == (unsigned int)m_bead.num_layers);
+    assert(d["Bead Section"]["Material ID"].Size() == (unsigned int)m_bead.num_layers);
     for (int i = 0; i < m_bead.num_layers; i++) {
         double thickness = d["Bead Section"]["Layer Thickness"][i].GetDouble();
         double angle = d["Bead Section"]["Ply Angle"][i].GetDouble();
@@ -106,8 +106,8 @@ void ANCFTire::ProcessJSON(const rapidjson::Document& d) {
     // Read layer information for the Sidewall Section
     m_sidewall.num_divs = d["Sidewall Section"]["Number Elements"].GetInt();
     m_sidewall.num_layers = d["Sidewall Section"]["Layer Thickness"].Size();
-    assert(d["Sidewall Section"]["Ply Angle"].Size() == m_sidewall.num_layers);
-    assert(d["Sidewall Section"]["Material ID"].Size() == m_sidewall.num_layers);
+    assert((int)d["Sidewall Section"]["Ply Angle"].Size() == m_sidewall.num_layers);
+    assert((int)d["Sidewall Section"]["Material ID"].Size() == m_sidewall.num_layers);
     for (int i = 0; i < m_sidewall.num_layers; i++) {
         double thickness = d["Sidewall Section"]["Layer Thickness"][i].GetDouble();
         double angle = d["Sidewall Section"]["Ply Angle"][i].GetDouble();
@@ -121,8 +121,8 @@ void ANCFTire::ProcessJSON(const rapidjson::Document& d) {
     // Read layer information for the Tread Section
     m_tread.num_divs = d["Tread Section"]["Number Elements"].GetInt();
     m_tread.num_layers = d["Tread Section"]["Layer Thickness"].Size();
-    assert(d["Tread Section"]["Ply Angle"].Size() == m_tread.num_layers);
-    assert(d["Tread Section"]["Material ID"].Size() == m_tread.num_layers);
+    assert((int)d["Tread Section"]["Ply Angle"].Size() == m_tread.num_layers);
+    assert((int)d["Tread Section"]["Material ID"].Size() == m_tread.num_layers);
     for (int i = 0; i < m_tread.num_layers; i++) {
         double thickness = d["Tread Section"]["Layer Thickness"][i].GetDouble();
         double angle = d["Tread Section"]["Ply Angle"][i].GetDouble();
